requirements.txt
requests==2.14.2
matplotlib==2.0.2
Pillow==4.1.1
colorama==0.3.9
pytesseract==0.1.7
baidu-aip==2.1.0.0
aircv==1.4.6
opencv-python==3.4.4.19